third_party/WebKit/LayoutTests/media/video-prefixed-fullscreen-expected.txt

CONSOLE WARNING: 'HTMLVideoElement.webkitSupportsFullscreen' is deprecated. Its value is true if the video is loaded.
CONSOLE WARNING: 'HTMLVideoElement.webkitDisplayingFullscreen' is deprecated. Please use the 'fullscreenchange' and 'webkitfullscreenchange' events instead.
CONSOLE WARNING: 'HTMLVideoElement.webkitEnterFullscreen()' is deprecated. Please use 'Element.requestFullscreen()' and 'Element.webkitRequestFullscreen()' instead.
CONSOLE WARNING: Failed to execute 'requestFullScreen' on 'Element': API can only be initiated by a user gesture.
CONSOLE WARNING: 'HTMLVideoElement.webkitExitFullscreen()' is deprecated. Please use 'Document.exitFullscreen()' and 'Document.webkitExitFullscreen()' instead.
Test the prefixed HTMLVideoElement fullscreen API

EXPECTED (video.webkitSupportsFullscreen == 'false') OK
EXPECTED (video.webkitDisplayingFullscreen == 'false') OK
EVENT(loadeddata)
EXPECTED (video.webkitSupportsFullscreen == 'true') OK
EXPECTED (video.webkitDisplayingFullscreen == 'false') OK
RUN(video.webkitEnterFullscreen())
EVENT(webkitfullscreenerror)
EXPECTED (video.webkitSupportsFullscreen == 'true') OK
EXPECTED (video.webkitDisplayingFullscreen == 'false') OK
RUN(video.webkitEnterFullscreen())
EVENT(webkitfullscreenchange)
EXPECTED (video.webkitSupportsFullscreen == 'true') OK
EXPECTED (video.webkitDisplayingFullscreen == 'true') OK
RUN(video.webkitExitFullscreen())
EVENT(webkitfullscreenchange)
EXPECTED (video.webkitSupportsFullscreen == 'true') OK
EXPECTED (video.webkitDisplayingFullscreen == 'false') OK
END OF TEST
